ADD: Class Reporte

+++ FinTec/Domain/CategoriaGasto.cs
@@ -6,11 +6,11 @@ namespace Domain
     {
         private Categoria _categoria;
         private double _montoUsado;
-        private int _porcentaje;
+        private double _porcentaje;
 
         public Categoria Categoria { get { return _categoria; } set { _categoria = value; } }
         public double MontoUsado { get { return _montoUsado; } set { _montoUsado = value;} }
-        public int Porcentaje { get { return _porcentaje; } set { _porcentaje = value; } }
+        public double Porcentaje { get { return _porcentaje; } set { _porcentaje = value; } }
 
         public CategoriaGasto(){}
 
@@ -21,7 +21,7 @@ namespace Domain
             _porcentaje = 0;
         }
 
-        public CategoriaGasto(Categoria c, double m, int p)
+        public CategoriaGasto(Categoria c, double m, double p)
         {
             _categoria= c;
             _montoUsado= m;
